% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/distCalc.R
\name{distCalc}
\alias{distCalc}
\title{Generate an inverse distance matrix for points within a landscape}
\usage{
distCalc(raster, coords, mat_type = "weight")
}
\arguments{
\item{raster}{a raster upon which distance are calculated}

\item{coords}{a file containing sample coordinates in lat long format}

\item{mat_type}{an character string, indicating whether the returned matrix should contain spatial "weight" or absolute "distance".}
}
\value{
a square matrix containing the inverse pairwise distance between all
points in \code{coords}.
}
\description{
Generate an inverse distance matrix for points within a landscape
}

