% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rasterWriter.R
\name{rasterWriter}
\alias{rasterWriter}
\title{Function for generating a raster from a shapefile}
\usage{
rasterWriter(shp, res, rastOut)
}
\arguments{
\item{shp}{a shape file to be converted to raster}

\item{res}{the resolution of the output raster}

\item{rastOut}{an string specifying the output name for the raster generated}
}
\value{
a raster file with the extension .asc
}
\description{
Function for generating a raster from a shapefile
}

